Django-Rest-02/requirements.txt
asgiref==3.5.2
astroid==2.4.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
Faker==14.2.0
gunicorn==20.1.0
idna==3.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
postgres==4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycodestyle==2.9.1
pylint==2.5.3
python-dateutil==2.8.2
pytz==2020.1
requests==2.28.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.12
validate-docbr==1.10.0
whitenoise==6.2.0
wrapt==1.12.1
